api/requirements.txt
adal==1.2.7
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
Authlib==1.3.2
bcrypt==4.0.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
cryptography==43.0.3
debugpy==1.8.8
decorator==5.1.1
distro==1.9.0
dnspython==2.4.2
email-validator==2.1.0.post1
executing==2.1.0
fastapi==0.115.0
fastapi-cli==0.0.4
fastapi-limiter==0.1.6
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identicons==0.1.0
idna==3.7
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.0
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
motor==3.6.0
msal==1.31.0
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.1.0
mypy==1.13.0
mypy-extensions==1.0.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.37.1
packaging==24.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.12
pyzmq==26.2.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
ruff==0.0.47
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.37.2
structlog==24.4.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
yarl==1.17.1